Fixing issue with Undefined Index
__METHOD__
I'm not really sure what caused this exactly, but an affected customer
had some notebooks being returned in linkedNotebooks and also business
shared notebooks but not business linked notebooks.

__EXPECTED__
The notebooks should still be returned.

__ACTUAL___
An undefined index error is coming up causing the script to just halt.

__CAUSE__
It's being assumed that if the linked notebook is in the shared business
notebooks, that it should also be in the linked business notebooks but
it's not for some unknown reason.

__FIX__
Since the item has been confirmed to have been shared because of the
shareKey comparison, making the business notebook item be null if the
linked business notebook entry is not found which will still have a
notebook instance which we can retrieve name and id information from.